backend/requirements.txt
flask
flask-cors
pandas
httpx
openai
langchain-openai
langchain-core
langchain-huggingface
langchain-pinecone
python-dotenv
pinecone-client
sentence_transformers
numpy<2